Vaadin components import WIP

+++ disconnect-polymer/src/main/java/js/lang/external/polymer/iron_iconset_svg/IronIconsetSvgElement.java
@@ -5,8 +5,8 @@ import com.github.fluorumlabs.disconnect.core.annotations.NpmPackage;
 import java.lang.String;
 import javax.annotation.Nullable;
 import js.lang.Any;
-import js.lang.external.polymer.legacy.LegacyElementMixin;
 import js.web.dom.Element;
+import js.web.dom.HTMLElement;
 import org.teavm.jso.JSProperty;
 
 /**
@@ -69,14 +69,14 @@ public interface IronIconsetSvgElement extends HTMLElement, LegacyElementMixin {
    *
    */
   @JSProperty("size")
-  int getSize();
+  double getSize();
 
   /**
    * The size of an individual icon. Note that icons must be square.
    *
    */
   @JSProperty("size")
-  void setSize(int value);
+  void setSize(double value);
 
   /**
    * Set to true to enable mirroring of icons where specified when they are
